+++ include/Hit.h
@@ -6,7 +6,7 @@ struct Hit
 {
     Vector position;
     Vector normal;
-    float distance = 0; // Distance from ray origin to hit position
+    double distance = 0; // Distance from ray origin to hit position
     size_t material = 0;
 
     Hit() = default;
