+++ include/RandomGenerator.h
@@ -7,9 +7,9 @@ class RandomGenerator
 public:
     RandomGenerator();
 
-    float value(float scale = 1.0f);
+    double value(double scale = 1.0f);
 
 private:
     std::mt19937 m_generator;
-    std::uniform_real_distribution<float> m_distribution{0.0f, 1.0f};
+    std::uniform_real_distribution<double> m_distribution{0.0, 1.0};
 };
