+++ src/Ray.cpp
@@ -15,17 +15,17 @@ Ray::Ray(Vector iorigin, Vector idirection)
 
 bool rayIntersectsBounds(const Ray& ray, const Bounds& bounds)
 {
-    float tmin = 0.0f;
-    float tmax = std::numeric_limits<float>::max();
+    double tmin = 0.0;
+    double tmax = std::numeric_limits<double>::max();
 
     for (int i = 0; i < 3; ++i)
     {
         const Axis axis = static_cast<Axis>(i);
         const Limits limits = bounds[axis];
-        const float origin = ray.origin[axis];
-        const float direction = ray.direction[axis];
+        const double origin = ray.origin[axis];
+        const double direction = ray.direction[axis];
 
-        if (std::abs(direction) < std::numeric_limits<float>::epsilon())
+        if (std::abs(direction) < std::numeric_limits<double>::epsilon())
         {
             if (origin < limits.min || origin > limits.max)
             {
@@ -34,9 +34,9 @@ bool rayIntersectsBounds(const Ray& ray, const Bounds& bounds)
         }
         else
         {
-            float ood = 1.0f / direction;
-            float t1 = (limits.min - origin) * ood;
-            float t2 = (limits.max - origin) * ood;
+            double ood = 1.0 / direction;
+            double t1 = (limits.min - origin) * ood;
+            double t2 = (limits.max - origin) * ood;
 
             if (t1 > t2)
             {
@@ -66,39 +66,39 @@ std::optional<Hit> rayIntersectsTriangle(const Ray& ray, const Triangle& triangl
 
     Vector n = Vector::cross(ab, ac);
 
-    float d = Vector::dot(qp, n);
+    double d = Vector::dot(qp, n);
 
-    if (d <= 0.0f)
+    if (d <= 0.0)
     {
         return std::nullopt;
     }
 
     Vector ap = ray.origin - triangle.a;
-    float t = Vector::dot(ap, n);
+    double t = Vector::dot(ap, n);
 
-    if (t < 0.0f)
+    if (t < 0.0)
     {
         return std::nullopt;
     }
 
     Vector e = Vector::cross(qp,ap);
-    float v = Vector::dot(ac, e);
-    if (v < 0.0f || v > d)
+    double v = Vector::dot(ac, e);
+    if (v < 0.0 || v > d)
     {
         return std::nullopt;
     }
 
-    float w = -Vector::dot(ab, e);
-    if (w < 0.0f || v + w > d)
+    double w = -Vector::dot(ab, e);
+    if (w < 0.0 || v + w > d)
     {
         return std::nullopt;
     }
 
-    float ood = 1.0f / d;
+    double ood = 1.0 / d;
     t *= ood;
     v *= ood;
     w *= ood;
-    float u = 1.0f - v - w;
+    double u = 1.0 - v - w;
 
     Vector coords{u, v, w};
     hit.position = triangle.getPosition(coords);
@@ -110,16 +110,16 @@ std::optional<Hit> rayIntersectsTriangle(const Ray& ray, const Triangle& triangl
 
 std::optional<Hit> rayIntersectsPlane(const Ray& ray, const Plane& plane)
 {
-    float dot = Vector::dot(plane.normal, ray.direction);
+    double dot = Vector::dot(plane.normal, ray.direction);
 
-    if (std::abs(dot) <= std::numeric_limits<float>::epsilon())
+    if (std::abs(dot) <= std::numeric_limits<double>::epsilon())
     {
         return std::nullopt;
     }
 
-    float t = (plane.dot - Vector::dot(plane.normal, ray.origin)) / dot;
+    double t = (plane.dot - Vector::dot(plane.normal, ray.origin)) / dot;
 
-    if (t < 0.0f)
+    if (t < 0.0)
     {
         return std::nullopt;
     }
